Added Missing Num
Added missing num. Another variation on the idea of arrays and finding
which number is missing. This works with an array of sequential numbers
in a random order where one is missing. Triangular Numbers is a key math
concept that makes this one a bit easier but there are other approaches
such as counting with the iterator and the array content to find the
missing value