Added valid anagram
Interesting one to check for anagrams. Letters in each string are
assigned to a corresponding slot in a size 26 array. For string 1
they're incremented by one, string 2 decrements by one. The result
should be zero if it is a valid anagram, the sum of the array is
insensitive to letter order