Added Single Number III
Wow this is an interesting one. The XOR of the array will produce the
XOR of the two desired numbers. To seperate the two we can then XOR the
array again but this time use a bit where the numbers differ and XOR was
1 to create to groups of numbers to XOR in our results array. Certainly
one to look at again in the future