Added strStr
This is a needle and haystack type of problem for a string within a
string. Reasonably straight forward, unforunately it is O(n^m)
complexity. Unfortunately no other data structures seem appropriate,
perhaps a linked list type of queue but once we have a match amongst
chars the queue would then be inaccruate when looking further downstream